fix: 修复在要备份的文件/文件夹不存在时，无法继续备份的问题

+++ ClassIsland/Services/FileFolderService.cs
@@ -188,11 +188,19 @@ public class FileFolderService(SettingsService settingsService, ILogger<FileFold
         {
             foreach (var i in backupFiles)
             {
+                if (!Path.Exists(Path.Combine(rootPath, i)))
+                {
+                    continue;
+                }
                 File.Copy(Path.Combine(rootPath, i), Path.Combine(backupTarget, i));
             }
 
             foreach (var i in backupFolders)
             {
+                if (!Path.Exists(Path.Combine(rootPath, i)))
+                {
+                    continue;
+                }
                 CopyFolder(Path.Combine(rootPath, i), Path.Combine(backupTarget, i));
             }
         });
